+++ ros2_ws/src/layer/control/control_interface/src/mode_control.cpp
@@ -11,7 +11,7 @@ ModeControl::ModeControl(){
 
 void ModeControl::unlock(
     uint8_t arm_mode,
-    uint8_t offboard_mode,
+    uint16_t offboard_mode,
     common_msgs::msg::ArmOffboardStatus current,
     common_msgs::msg::ArmOffboardStatus &pub_msg)
 {   
@@ -39,7 +39,7 @@ void ModeControl::locked(
 
 void ModeControl::set_mode(
     uint8_t target_arm,
-    uint8_t target_offboard,
+    uint16_t target_offboard,
     const common_msgs::msg::ArmOffboardStatus &current,
     common_msgs::msg::ArmOffboardStatus &pub_msg,
     const std::string &action_desc)
